sql: move code related to casts into separate file
Release note: None
